refactor(api-file-manager): migrate graphql layer to GraphQLSchemaFactory DI
Replace GraphQLSchemaPlugin + ContextPlugin pattern with GraphQLSchemaFactory
implementations. Schema factories receive dependencies via constructor injection
and use builder.addResolver with DI-resolved deps instead of
context.container.resolve() in each resolver.

- FmBaseSchema: settings query/mutation
- FmFilesSchema: file CRUD + dynamic CMS model typedefs
- FmGetFileByUrlSchema: file lookup by URL
- createFilesTypeDefs inlined into FmFilesSchema
- createFileManagerGraphQL() returns empty array (schemas auto-discovered via DI)
