--- packages/api-file-manager/src/graphql/utils.ts
@@ -1,1 +1,0 @@
-export const emptyResolver = () => ({});
